# Optimize model encoding