--- a/.gitignore
+++ b/.gitignore
@@ -1,1 +1,3 @@
 *.swp
+*.test
+*.out