… internal/fastjson/LICENSE.easyjson.txt
… internal/fastjson/generate.go
… model/marshal_fastjson.go
… model/marshal_test.go