- simplify the model types, and push encoding to the tracer. e.g. timestamp is now a string, duration is a number; the tracer is responsible for formatting timestamps (using the provided model.FormatTime) and durations (duration.Seconds() * 1000).

- introduce a new internal package, fastjson. I started out using easyjson, but wasn't happy with the code it produced. I created a smaller, more limited code generator that does what apm-agent-go needs, while producing significantly better code for the use case. Since we control all encoding sites, and there's one goroutine that does all encoding, we can reuse buffers without going through a sync.Pool.

- Some other micro optimisations, e.g. introducing apmhttp.StatusCodeString for mapping status code integers to strings, to avoid having to call into strconv.Itoa.

I micro-benchmarked the changes compared to the previous commit, copying across BenchmarkMarshalTransactionPayload, changing the test to encode with encoding/json.Marshal as the tracer did before.

```
benchmark                                      old ns/op     new ns/op
delta
BenchmarkMarshalTransactionPayloadStdlib-4     28949952      1649420
-94.30%

benchmark                                      old allocs     new allocs
delta
BenchmarkMarshalTransactionPayloadStdlib-4     66014          0
-100.00%

benchmark                                      old bytes     new bytes
delta
BenchmarkMarshalTransactionPayloadStdlib-4     10023684      5879
-99.94%
```